Dead Man switch smart contract
Dead Man's Switch 💀
A dead man's switch is a term used in many cases. The idea is pretty simple: if someone should become incapacitated, there is a way to detect it and act on it.

For our particular case we'll create a mechanism where the owner of a contract will need to ping or notify a contract every 52 weeks. If there is no activity during this time period, the recipient will be able to withdraw the funds. The assumption here is that the owner is no longer able to do so.

🔒 Contract Security

Ensure that only the owner can call ping. No other address should be able to delay the switch's withdrawal execution. Revert the transaction if the caller is any other address.

Ensure that withdraw can only be called after 52 weeks of inactivity. If the owner has called ping or deployed the contract more recently than that, the withdrawal should be reverted.